Move Top-N metrics around
